+++ focWebServer/src/com/foc/vaadin/gui/tableExports/FVTableToCSVExport.java
@@ -26,8 +26,8 @@ public class FVTableToCSVExport extends FVTableTreeCSVExport {
     Iterator localIterator = localCollection.iterator();
     while ((localIterator != null) && (localIterator.hasNext()))
     {
-      Integer localInteger = (Integer)localIterator.next();
-      FocObject localFocObject = localFocList.searchByReference(localInteger.intValue());
+      Long localInteger = (Long)localIterator.next();
+      FocObject localFocObject = localFocList.searchByReference(localInteger.longValue());
       if (localFocObject != null)
       {
         for (int i = 0; i < localArrayList.size(); i++)
