+++ focWebServer/src/com/foc/vaadin/gui/tableExports/FVTableTreeCSVExport.java
@@ -153,7 +153,7 @@ public abstract class FVTableTreeCSVExport extends CSVExport {
 					if((localObject1 instanceof Table.ColumnGenerator)){
 						Table.ColumnGenerator localColumnGenerator = (Table.ColumnGenerator) localObject1;
 						if(localColumnGenerator != null){
-							Integer itemID = Integer.valueOf(paramFocObject.getReference().getInteger());
+							Long itemID = Long.valueOf(paramFocObject.getReference().getLong());
 							localObject2 = localColumnGenerator.generateCell(getTableTreeDelegate().getTable(), itemID, paramFVTableColumn.getName());
 							if(localObject2 != null){
 								if((localObject2 instanceof FVMultipleChoiceOptionGroupPopupView)){
